chore: rename tooling => hack
