+++ hack/crdoc/markdown.tmpl
