+++ hack/scripts/annotate-crds.mjs
