+++ hack/scripts/get-chart-version.mjs
