+++ hack/scripts/lib/index.mjs
