+++ hack/scripts/lib/version.mjs
