+++ sonar-project.properties
@@ -4,7 +4,7 @@ sonar.organization=gravitee-io
 sonar.host.url=https://sonarcloud.io
 
 sonar.sources=.
-sonar.exclusions=test/**/*,tooling/scripts/**/*,tooling/kind/**/*,examples/**/*,**/zz_generated.deepcopy.go,main.go
+sonar.exclusions=test/**/*,hack/scripts/**/*,hack/kind/**/*,examples/**/*,**/zz_generated.deepcopy.go,main.go
 
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
